Step 4: P balanced corrected! (ASM2d-ADM1)
In Step 4, when updating P in Case II where N is the limiting case, we need to update X_S_P only by the amount of Protein change in that part which is bio2pr, not X_pr. This is because X_pr has already been formed in Step 3.

Additionally, in case III when limiting component is P we cannot simply decide the limiting case, because X_pr_i_N >> X_pr_i_P which form the denominator. Hence two cases within case III.

The first of these two changes led to TP balance for METRO influent. WOOT!