Step 6: N balance holds! (ASM2d-ADM1)
In Step 6, The original ASM1-ADM1 implementation never accounted for N present in soluble inert component because ASM1 (S_I_i_N) = 0, but ASM2d (S_I_i_N) = 0.01. This resulted in an error equal to the N content in S_I. Now we are accounting for N in S_I and thus, this issue is resolved. N is balanced in the developed interface. WOOT!