Correct checks for modified ASM2d-ADM1
X_S in ASM1 did not have N content, as it was taken care of by X_ND. However, in ASM2d X_S has non-zero N content, which we use as 'X_ND_asm1' in the code. Therefore, we no longer need the check (same with S_F). On the other hand N/P content in S_A now needs a check.